# openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Questerra — player API",
        "version": "v1",
        "description": "Generated by tools/openapi-gen.php from the running router, the controllers'\nvalidate() rules, and a live probe of GET responses. Do not edit by hand: the\nhand-written contract drifted, which is why this file exists.\n\nSemantics a schema cannot carry — that submit stays open after time_expired, that\nfun_game answers score zero, the minimum interval between position posts — live in\nAPK-BUILD-GUIDE.md and still have to be read.",
        "x-behaviour-guides": {
            "API-V1-CONTRACT.md": "8afbdf4e5e3c2c79",
            "APK-BUILD-GUIDE.md": "a83f14c3f7f0267b",
            "APK-WORK-ORDER.md": "927669638c184ad6"
        }
    },
    "servers": [
        {
            "url": "https://questerra-series.com"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Laravel Sanctum personal access token"
            }
        }
    },
    "paths": {
...
        "/api/v1/ar/locations/{id}": {
            "get": {
                "operationId": "api.v1.ar.location",
                "summary": "api.v1.ar.location",
                "x-auth": "sanctum-bearer",
                "x-access-window": true,
                "x-rate-limit": "api",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Not probed — POST endpoints are not called by the generator"
                    }
                }
            }
        },
...
        "/api/v1/features": {
            "get": {
                "operationId": "api.v1.features",
                "summary": "api.v1.features",
                "x-auth": "sanctum-bearer",
                "x-access-window": true,
                "x-rate-limit": "api",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Observed on a live probe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "features": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "enabled": {
                                                        "type": "boolean"
                                                    },
                                                    "key": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "sort_order": {
                                                        "type": "integer"
                                                    }
                                                }
                                            }
                                        },
                                        "flags": {
                                            "type": "object",
                                            "properties": {
                                                "dashboard_quick_stat_average_score": {
                                                    "type": "boolean"
                                                },
                                                "dashboard_quick_stat_completed": {
                                                    "type": "boolean"
                                                },
                                                "dashboard_quick_stat_streak": {
                                                    "type": "boolean"
                                                },
                                                "dashboard_quick_stat_total_time": {
                                                    "type": "boolean"
                                                },
                                                "dashboard_quick_stats": {
                                                    "type": "boolean"
                                                },
                                                "dashboard_stat_average_points": {
                                                    "type": "boolean"
                                                },
                                                "dashboard_stat_completed": {
                                                    "type": "boolean"
                                                },
                                                "dashboard_stat_completion_rate": {
                                                    "type": "boolean"
                                                },
                                                "dashboard_stat_history": {
                                                    "type": "boolean"
                                                },
                                                "dashboard_stat_total_attempts": {
                                                    "type": "boolean"
                                                },
                                                "dashboard_stat_total_score": {
                                                    "type": "boolean"
                                                },
                                                "game_dashboard": {
                                                    "type": "boolean"
                                                },
                                                "gps_tracking": {
                                                    "type": "boolean"
                                                },
                                                "leaderboard": {
                                                    "type": "boolean"
                                                },
                                                "notifications": {
                                                    "type": "boolean"
                                                },
                                                "quest_locations": {
                                                    "type": "boolean"
                                                },
                                                "quiz_system": {
                                                    "type": "boolean"
                                                },
                                                "team_management": {
                                                    "type": "boolean"
                                                },
                                                "user_dashboard_session_timer": {
                                                    "type": "boolean"
                                                },
                                                "user_dashboard_stats": {
                                                    "type": "boolean"
                                                },
                                                "workflow_timers": {
                                                    "type": "boolean"
                                                }
                                            }
                                        },
                                        "success": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/indoor-map/{id?}": {
            "get": {
                "operationId": "api.v1.indoor-map",
                "summary": "api.v1.indoor-map",
                "x-auth": "sanctum-bearer",
                "x-access-window": true,
                "x-rate-limit": "api",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Not probed — POST endpoints are not called by the generator"
                    }
                }
            }
        },
...
        "/api/v1/quest-locations/checkin": {
            "post": {
                "operationId": "api.v1.quest.checkin",
                "summary": "api.v1.quest.checkin",
                "x-auth": "sanctum-bearer",
                "x-access-window": true,
                "x-rate-limit": "api",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "accuracy": {
                                        "nullable": true,
                                        "type": "number",
                                        "minimum": 0,
                                        "x-laravel-rules": "nullable|numeric|min:0"
                                    },
                                    "device_info": {
                                        "nullable": true,
                                        "type": "string",
                                        "maxLength": 255,
                                        "x-laravel-rules": "nullable|string|max:255"
                                    },
                                    "location_id": {
                                        "type": "integer",
                                        "description": "Must already exist: quest_locations,id",
                                        "x-laravel-rules": "required|exists:quest_locations,id"
                                    },
                                    "user_latitude": {
                                        "type": "number",
                                        "x-laravel-rules": "required|numeric"
                                    },
                                    "user_longitude": {
                                        "type": "number",
                                        "x-laravel-rules": "required|numeric"
                                    }
                                },
                                "required": [
                                    "location_id",
                                    "user_latitude",
                                    "user_longitude"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Not probed — POST endpoints are not called by the generator"
                    }
                }
            }
        },
...
        "/api/v1/quiz/complete-game": {
            "post": {
                "operationId": "api.v1.quiz.complete-game",
                "summary": "api.v1.quiz.complete-game",
                "x-auth": "sanctum-bearer",
                "x-access-window": true,
                "x-rate-limit": "api",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "attempt_id": {
                                        "type": "integer",
                                        "x-laravel-rules": "required|integer"
                                    },
                                    "question_id": {
                                        "type": "integer",
                                        "x-laravel-rules": "required|integer"
                                    }
                                },
                                "required": [
                                    "attempt_id",
                                    "question_id"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Not probed — POST endpoints are not called by the generator"
                    }
                }
            }
        },
...
    }
}